+++ postgres/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Add `load_balance_hosts` to `Config` struct.
+- Add `hostaddr` and `hostaddrs` to `Config` struct.
 - Add `url` field to `Config` struct. This enables parsing of
   connection URLs.